Some of the connectors were not inheriting from AbstractSpoolDirSourceConnector so they did not support multiple tasks. Moved this functionality to AbstractSourceConnector. Fixes #136
