+++ src/main/java/com/github/jcustenborder/kafka/connect/spooldir/SpoolDirCsvSourceConnector.java
@@ -35,7 +35,7 @@ import java.util.Map;
     "")
 public class SpoolDirCsvSourceConnector extends AbstractSpoolDirSourceConnector<SpoolDirCsvSourceConnectorConfig> {
   @Override
-  protected SpoolDirCsvSourceConnectorConfig config(Map<String, String> settings) {
+  protected SpoolDirCsvSourceConnectorConfig config(Map<String, ?> settings) {
     return new SpoolDirCsvSourceConnectorConfig(false, settings);
   }
