+++ src/main/java/com/github/jcustenborder/kafka/connect/spooldir/SpoolDirSchemaLessJsonSourceConnector.java
@@ -15,17 +15,12 @@
  */
 package com.github.jcustenborder.kafka.connect.spooldir;
 
-import com.github.jcustenborder.kafka.connect.utils.VersionUtil;
 import com.github.jcustenborder.kafka.connect.utils.config.Description;
 import com.github.jcustenborder.kafka.connect.utils.config.DocumentationImportant;
 import com.github.jcustenborder.kafka.connect.utils.config.Title;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
-import org.apache.kafka.connect.source.SourceConnector;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 @Title("Schema Less Json Source Connector")
@@ -34,14 +29,10 @@ import java.util.Map;
 @DocumentationImportant("This connector does not try to convert the json records to a schema. " +
     "The recommended converter to use is the StringConverter. " +
     "Example: `value.converter=org.apache.kafka.connect.storage.StringConverter`")
-public class SpoolDirSchemaLessJsonSourceConnector extends SourceConnector {
-
-  Map<String, String> settings;
-
+public class SpoolDirSchemaLessJsonSourceConnector extends AbstractSourceConnector<SpoolDirSchemaLessJsonSourceConnectorConfig> {
   @Override
-  public void start(Map<String, String> settings) {
-    SpoolDirSchemaLessJsonSourceConnectorConfig config = new SpoolDirSchemaLessJsonSourceConnectorConfig(settings);
-    this.settings = settings;
+  protected SpoolDirSchemaLessJsonSourceConnectorConfig config(Map<String, ?> settings) {
+    return new SpoolDirSchemaLessJsonSourceConnectorConfig(settings);
   }
 
   @Override
@@ -50,31 +41,7 @@ public class SpoolDirSchemaLessJsonSourceConnector extends SourceConnector {
   }
 
   @Override
-  public List<Map<String, String>> taskConfigs(int taskCount) {
-    List<Map<String, String>> result = new ArrayList<>();
-
-    for (int i = 0; i < taskCount; i++) {
-      Map<String, String> taskConfig = new LinkedHashMap<>(this.settings);
-      taskConfig.put(AbstractSourceConnectorConfig.TASK_INDEX_CONF, Integer.toString(i));
-      taskConfig.put(AbstractSourceConnectorConfig.TASK_COUNT_CONF, Integer.toString(taskCount));
-      result.add(taskConfig);
-    }
-
-    return result;
-  }
-
-  @Override
-  public void stop() {
-
-  }
-
-  @Override
   public ConfigDef config() {
     return SpoolDirSchemaLessJsonSourceConnectorConfig.config();
-  }
-
-  @Override
-  public String version() {
-    return VersionUtil.version(this.getClass());
   }
 }
